Now able to calculate taxes. Need to add more optional parameters to get a more accurate calculation.

+++ TaxJar.Web/Models/Requests/TaxJarCalulateRequest.cs
@@ -41,7 +41,7 @@ namespace TaxJar.Api.Models.Requests
         public int Amount;
 
         [JsonProperty("shipping")]
-        public double Shipping;
+        public float Shipping;
 
         [JsonProperty("nexus_addresses")]
         public List<NexusAddress> NexusAddresses;
